move command context to specific file
